Create largestdigitcombo.cpp
You are given n integers. Your task is to create an integer X which has n digits in total. The i-th digit of integer X should be the largest digit among the digits of the i-th integer.



Example:



Input:

3

4312

5678

1234



Output:

484



Explanation:

For the first integer 4312, the largest digit is 4.
For the second integer 5678, the largest digit is 8.
For the third integer 1234, the largest digit is 4.
Thus, the resulting integer X is 484.

Procedure: You must follow the instructions while coding

giveLargest function
Should have one argument, an integer n
Accept n numbers in the function
Perform the computation as mentioned above and create the number X
Return this number X to the main function
Main Program: 
Accept an integer n from the user and pass it to the function giveLargest. 
Print the result passed by the function giveLargest into the main program


Input Format:
The first line contains an integer n (1 <= n <= 100), representing the number of integers.
The next n lines each contain an integer, where each integer will be in the range [1, 99999999].


Output Format:
Output a single integer X, which consists of n digits. The i-th digit of X is the largest digit from the i-th integer.