Create decodingciper.cpp
You are given a string containing a mix of encoded cipher types. The string is encoded using one or more of the following cipher types:

Number-to-Letter Cipher (Type 1): The number from 1 to 26 represents the corresponding letter in the alphabet, where 1 = A, 2 = B, ..., 26 = Z.
Shifted Caesar Cipher (Type 2): Character is shifted by a constant amount k. You need to decode the character by shifting each letter back by k positions. For example, if k = 3, 'A' becomes 'X', 'B' becomes 'Y', etc.
Repeating Letter Cipher (Type 3): The encoded string contains a number that represents how many times a letter is repeated. For example, 3A means the letter 'A' repeated three times (AAA).
You need to decode the entire encoded message. The cipher types will alternate in the following order:

Type 1 (Number-to-Letter Cipher)
Type 2 (Shifted Caesar Cipher)
Type 1 (Number-to-Letter Cipher)
Type 3 (Repeating Letter Cipher)
And so on...
Your task is to decode the entire string.

Example:

Input:

5

1 8

3 A 3

1 13

2 B 2

3 C 5



Output:

HAAAMZCCCCC

Note: Your output may print line by line, i.e. when you input 1 8 for the example above, H may be printed at the same time, which is fine, you don't need to use strings for this question

Input Format:
The first line contains an integer n (1 <= n <= 100), the number of encoding operations.
In the next n lines, first will be a integer representing the type of cipher and then the :
For type 1, First integer will denote the type of cipher (i.e. 1), the next integer will correspond to a letter.
For type 2, First integer will denote the type of cipher (i.e. 2), the next will be a character and then there will be an integer.
For type 3, First integer will denote the type of cipher (i.e. 3), the next will be a character and then there will be an integer.
Output Format:
Print the encoded cipher.


Assumptions on input:
Characters will always be CAPITALIZED.


Note:
Do not write any C++ statements for printing general messages. For example, the following should NOT be present in your program:
cout << "Enter a number:",  
cout << "The computed answer is", etc.
cout should be used to print only the computed final output. In addition, do not print unnecessary spaces unless specified in the program.
If any hard coding is found, or if any test case passes by merely writing a cout statement and without any logic, then the marks for that test case will NOT be awarded.